Split events over multiple handlers, switched to Serilog
